+++ lvlgen.py
@@ -1,11 +1,10 @@
 import pygame as pg
-import math, time, os
-from pygame import gfxdraw
+import os
 
 path = os.getcwd()
 
 #Game build associated with level generator
-buildId = "id152.1"
+buildId = "id166.1"
 
 def saveARL(ls,ls2,dest):
     bitO = []